+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.6.0
-* Move configuration in a specific window
+* Move configuration in a specific window, and allow dynamic modification of config files
 * Add visualization of to-be-processed inputs while highlighting the currently processed input, with a dedicated progress bar 
 * Add parameters for pre-mask method and and branch signal correction
 * Take config into account to validate/invalidate cache in debug mode